Prva verzija parsera.

+++ hw03-1191228748/src/main/java/hr/fer/zemris/java/custom/scripting/elems/ElementConstantDouble.java
@@ -21,7 +21,7 @@ public class ElementConstantDouble extends Element {
 	 * 
 	 * @param value value of the double
 	 */
-	public ElementConstantDouble(int value) {
+	public ElementConstantDouble(double value) {
 		this.value = value;
 	}
